# ENH: Extend np.add ufunc to work with unicode and byte dtypes