--- a/doc/release/upcoming_changes/24858.new_feature.rst
+++ b/doc/release/upcoming_changes/24858.new_feature.rst
@@ -1,0 +1,1 @@
+Extend ``np.add`` to work with ``unicode`` and ``bytes`` dtypes.