--- a/numpy/core/tests/test_defchararray.py
+++ b/numpy/core/tests/test_defchararray.py
@@ -678,15 +678,3 @@ def test_empty_indexing():
     # empty chararray instead of a chararray with a single empty string in it.
     s = np.char.chararray((4,))
     assert_(s[[]].size == 0)
-
-
-@pytest.mark.parametrize(["dt1", "dt2"],
-        [("S", "U"), ("U", "S"), ("S", "O"), ("U", "O"),
-         ("S", "d"), ("S", "V")])
-def test_add_types(dt1, dt2):
-    arr1 = np.array([1234234], dtype=dt1)
-    # If the following fails, e.g. use a number and test "V" explicitly
-    arr2 = np.array([b"423"], dtype=dt2)
-    with pytest.raises(TypeError,
-            match=f".*same dtype kind.*{arr1.dtype}.*{arr2.dtype}"):
-        np.char.add(arr1, arr2)