--- a/numpy/core/tests/test_ufunc.py
+++ b/numpy/core/tests/test_ufunc.py
@@ -2878,7 +2878,8 @@ def test_addition_string_types(dt1, dt2):
         np.add(arr1, arr2)
 
 
-@pytest.mark.parametrize("order1,order2", [(">", ">"), ("<", "<"), (">", "<"), ("<", ">")])
+@pytest.mark.parametrize("order1,order2",
+                         [(">", ">"), ("<", "<"), (">", "<"), ("<", ">")])
 def test_addition_unicode_inverse_byte_order(order1, order2):
     element = 'abcd'
     arr1 = np.array([element], dtype=f"{order1}U4")