@lysnikolaou 2023-10-05
One question to ask ourselves here is whether we want to also implement loops for mixed dtypes like `unicode + bytes`, `unicode + object`, `bytes + object`, `unicode + void` and `bytes + void`.