@mattip 2023-10-05
I don't think the ufunc fastpath has anything to offer once `object` dtype is used. And I don't know how common it is to add `string_` and `bytes_`, so my first instinct would be to keep this as simple as possible.