@lysnikolaou 2023-10-05
Here's another run of the benchmarks.

```
| Change   | Before [64fc516a] <main>   | After [6b970817] <string-ufuncs-add>   |   Ratio | Benchmark (Parameter)                               |
|----------|----------------------------|----------------------------------------|---------|-----------------------------------------------------|
| -        | 18.0±0.3μs                 | 3.04±0.05μs                            |    0.17 | bench_core.NumPyChar.time_add_small_list_big_string |
| -        | 3.29±0.01ms                | 9.77±0.5μs                             |    0    | bench_core.NumPyChar.time_add_big_list_small_string |
```