@ngoldbaum 2023-10-05
> And I don't know how common it is to add string_ and bytes_, so my first instinct would be to keep this as simple as possible.

I agree, we shouldn't assume the encoding of `bytes_` data, which you would need to do I think.