@charris 2023-10-16
It needs a release note. Look in `doc/release/upcoming_changes/` for examples.