@lysnikolaou 2023-10-16
@charris Added a minimal release note. Should it be bigger than that?